% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dscan.R
\name{deletelines}
\alias{deletelines}
\title{Clustering results filtering}
\usage{
deletelines(x, cutoff, colname = NULL)
}
\arguments{
\item{x}{data frame contain clustering results}

\item{cutoff}{the minimum number of spots contained in the cluster to be filtered}

\item{colname}{the column name of clustering result ID in x}
}
\value{
Clustering results after filtering
}
\description{
Filter cluster accoring to the number of cell in clusters.
}
