% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dscan.R
\name{extractCluster}
\alias{extractCluster}
\title{Draw the outline with convex hull method}
\usage{
extractCluster(x, subcluster, e = NULL)
}
\arguments{
\item{x}{data frame contain the cluster id and coordinates of cells named cluster,row,col}

\item{subcluster}{cluster id}

\item{e}{the expanding radius}
}
\value{
a list outer contour points contains two formats such as polygen and data frame
}
\description{
Get coordinates of polygon point for specific cluster with convex hull method,return a list contains two formats such as polygen and data frame.
}
