% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dscan.R
\name{extractCluster2}
\alias{extractCluster2}
\title{Draw the outline with alpha hull (a-shape) method}
\usage{
extractCluster2(x, cluster, e, alpha)
}
\arguments{
\item{x}{data frame contain the cluster id and coordinates of cells named cluster,row,col}

\item{cluster}{cluster id}

\item{e}{the expanding radius}

\item{alpha}{Value of α}
}
\value{
a list of outer contour points contains two formats such as polygen and data frame
}
\description{
Get coordinates of polygon point for specific cluster with alpha hull (a-shape) method,return a list contains two formats such as polygen and data frame.
}
