% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dscan.R
\name{point_in_polygon}
\alias{point_in_polygon}
\title{Get coordinates of spots within specific cluster}
\usage{
point_in_polygon(polygon, point)
}
\arguments{
\item{polygon}{polygon of cluster}

\item{point}{coordinates of spots}
}
\value{
coordinates of spots within specific cluster
}
\description{
Get coordinates of spots within specific cluster
}
