% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dscan.R
\name{rangexy}
\alias{rangexy}
\title{Get coordinates of spots within specific cluster}
\usage{
rangexy(frame, annofile, n, p)
}
\arguments{
\item{frame}{a data frame that contains the point coordinates for each cluster polygon. The row and col columns will contain the point coordinates as rows and columns, respectively, while the cluster column will indicate the cluster ID to which each point belongs}

\item{annofile}{a data frame of density clustering result,contains at least 3 columns:coordinates(row,col) and cluster id(cluster)}

\item{n}{specific cluster id}

\item{p}{index of specific cluster in polygon list}
}
\value{
a data frame contains coordinates of spots within specific cluster
}
\description{
Get coordinates of spots within specific cluster
}
