% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesetscore.R
\name{sc.Meta}
\alias{sc.Meta}
\title{Compute the score of specific geneset on single cell level with 9 optional methods}
\usage{
sc.Meta(
  object,
  assays = NULL,
  signaturesFile,
  method,
  imputation = FALSE,
  ncores = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{assays}{name of assays to store gene matrix}

\item{signaturesFile}{gmt file of geneset}

\item{method}{name of calculation method}

\item{imputation}{whether to use normalized data,default = FALSE}

\item{ncores}{number of cores used}
}
\value{
a data frame with module scores; each module is stored as name# for each module program present in features; 1 cell per row,1 module per column
}
\description{
Calculate the expression levels of each program (geneset) on single cell level with 9 optional methods such as AUCell, ssGSEA, gsva, AddModuleScore, mean, pagoda2, VAM, singscore, UCell.
}
