% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dscan.R
\name{score.subset}
\alias{score.subset}
\title{Select cell according to gene set score}
\usage{
score.subset(metaall, genename, cutoff, absolute = TRUE)
}
\arguments{
\item{metaall}{dataframe contain gene set score}

\item{genename}{name of gene set stored in the column names of the metaall}

\item{cutoff}{value of gene set score to cutoff}

\item{absolute}{use absolute values(TRUE) or percentages(FALSE),default = FALSE}
}
\value{
meta data frame of filtered spots
}
\description{
Select cell according to gene set score
}
