% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dscan.R
\name{subsetCluster}
\alias{subsetCluster}
\title{Specific cluster selecting and density clustering}
\usage{
subsetCluster(fx, clusterID, k, eps)
}
\arguments{
\item{fx}{data frame contain clustering results}

\item{clusterID}{ID of cluster to select}

\item{k}{the minimum number of points (a threshold) clustered together for a region to be considered dense}

\item{eps}{distance measure that will be used to locate the points in the neighborhood of any point}
}
\value{
data frame contains the result of a specific cluster reclustering
}
\description{
Select a specific cluster and perform density clustering.
}
