% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.Mtx2SeuratObj.R
\name{Transform.Mtx2SeuratObj}
\alias{Transform.Mtx2SeuratObj}
\title{Create a Spatial Seurat object from file generated with the Visium technology such as Stereo-seq}
\usage{
Transform.Mtx2SeuratObj(
  infile,
  bin.size = NULL,
  outdir,
  outfile,
  ncores = NULL
)
}
\arguments{
\item{infile}{file contains at least 4 columns such as the coordinates of spots(named x,y),geneID and UMICount}

\item{bin.size}{number of bins a spot contains}

\item{outdir}{output directory}

\item{outfile}{name of output file}

\item{ncores}{number of cores used}
}
\value{
a Spatial Seurat object
}
\description{
Create a Spatial Seurat object from file generated with the Visium technology such as Stereo-seq
}
